rename studioffoer
